% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arial.widths.r
\docType{data}
\name{arial.widths}
\alias{arial.widths}
\title{relative widths of characters for Arial font from 
https://gist.github.com/aminnj/5ca372aa2def72fb017b531c894afdca
Used by pubtheme to automatically position legends}
\format{
An object of class \code{data.frame} with 94 rows and 2 columns.
}
\usage{
arial.widths
}
\description{
relative widths of characters for Arial font from 
https://gist.github.com/aminnj/5ca372aa2def72fb017b531c894afdca
Used by pubtheme to automatically position legends
}
\keyword{datasets}
