% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{cb.pal}
\alias{cb.pal}
\title{colorblind friendly color palette for `pubtheme`}
\format{
An object of class \code{character} of length 14.
}
\usage{
cb.pal
}
\description{
colorblind friendly color palette for `pubtheme`
}
\keyword{datasets}
