% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{cb14}
\alias{cb14}
\title{colorblind friendly color palette}
\format{
An object of class \code{character} of length 15.
}
\source{
https://jacksonlab.agronomy.wisc.edu/2016/05/23/15-level-colorblind-friendly-palette/
}
\usage{
cb14
}
\description{
colorblind friendly color palette
}
\keyword{datasets}
