% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{default.pal}
\alias{default.pal}
\title{A color used in `pubtheme`}
\format{
An object of class \code{character} of length 7.
}
\usage{
default.pal
}
\description{
A color used in `pubtheme`
}
\keyword{datasets}
