% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.r
\name{ht}
\alias{ht}
\title{ht: quickly show both `head()` and `tail()` for a data.frame}
\usage{
ht(x, n = 2, keepnums = FALSE, ...)
}
\arguments{
\item{x}{An R object, usually a data frame, as in `head`.}

\item{n}{An integer vector as in `head`.}

\item{keepnums}{Same as in `head`.}
}
\value{
When used with a data frame, returns the first `n` and last `n` rows.
}
\description{
A convenience function that shows by default both the `head` and `tail` of a data frame or other objects supported by `head` and `tail`.
}
\examples{
ht(mtcars)
}
