% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutpub.r
\name{layoutpub}
\alias{layoutpub}
\title{A plotly theme that matches the ggplot theme theme_pub}
\usage{
layoutpub(
  p,
  type = "scatter",
  base_size = 18,
  subtitle = F,
  caption = F,
  legend.rows = 0,
  facet = F
)
}
\arguments{
\item{p}{A plotly object created using plot_ly. Future versions will allow plotly objects created using `ggplotly`.}

\item{type}{Text indicating the type of plot the theme is being used for.  Supported types are currently 'line', 'bar', 'hist', 'grid', 'scatter'. Other plot types will be added later.}

\item{base_size}{base font size, given in pts. For viewing in RStudio, something like 18 (the default) or 12 is recommended.}

\item{subtitle}{Is there a subtitle? Default is FALSE. This affects spacing.}

\item{caption}{Is there a caption? Default is FALSE. This affects spacing.}

\item{legend.rows}{How many rows are needed for the legend? Default is 0. This affects spacing.}

\item{facet}{Not implemented yet. Indicates whether or not `facet_wrap` or `facet_grid` are being used for this plot.  The default is `facet=FALSE`.}
}
\value{
When used in conjunction with plot_ly, it returns a plot formatted using layout.pub.
}
\description{
A plotly theme for making publication quality interactive visualizations with options for organization-specific colors and logos.
}
\examples{
#See `https://github.com/bmacGTPM/pubtheme` for examples.

}
