% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketitle.r
\name{maketitle}
\alias{maketitle}
\title{A function used by layout.pub for formating a title with optional subtitle.}
\usage{
maketitle(title, subtitle = NULL, base_size = 18)
}
\arguments{
\item{title}{Text with the desired title in Upper Lower.}

\item{subtitle}{Text with an optional subtitle in Upper Lower. Default is NULL.}

\item{base_size}{Font size. Default is 18. Should be set to the font size being used in `layout.pub`.}
}
\value{
Returns an HTML-formatted title (and optional subtitle) for use by layout.pub and plot_ly.
}
\description{
This function is a way to add a subtitle with lighter gray font color to a figure using layout.pub
}
\examples{
#See `https://github.com/bmacGTPM/pubtheme` for examples.
}
