% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.density.r
\name{pub.density}
\alias{pub.density}
\title{A density plot used in `pubtheme`}
\usage{
pub.density(data, mapping, ...)
}
\arguments{
\item{data}{The `data` as in `ggplot`}

\item{mapping}{The `mapping` as in `ggplot`}
}
\value{
When used in conjunction with ggplot, it returns a plot formatted using theme_pub.
}
\description{
A density plot that is a modified version of GGally::ggally_density that has more transparent filled areas under the density lines.
}
\examples{
#See the `Pairs Plot` section of `https://github.com/bmacGTPM/pubtheme` for examples.
}
