% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.r
\name{pub}
\alias{pub}
\title{Prepare plot for publication}
\usage{
pub(
  g,
  type = "scatter",
  base_size = 36/3,
  xlim = NULL,
  ylim = NULL,
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xtrans = NULL,
  ytrans = NULL,
  int = FALSE,
  tooltip = "all",
  subtitle = F,
  caption = F,
  legend.rows = 0,
  ...
)
}
\arguments{
\item{g}{A `ggplot` object.}

\item{type}{Text indicating the type of plot the theme is being used for.  Supported types are currently 'line', 'bar', 'hist', 'grid', 'scatter' which can be used for plots with `geom_line`, `geom_bar`, `geo_histogram`, `geom_tile`, and `geom_point`.  Other plot types will be added later.}

\item{base_size}{base font size, given in pts. For saving images, 36 is required. For viewing in RStudio, something much smaller like 36/3 = 12 (the default) is recommended.}

\item{xlim}{(Optional) A vector of length two, specifying the lower and upper limit of the x-axis}

\item{ylim}{(Optional) A vector of length two, specifying the lower and upper limit of the y-axis}

\item{xbreaks}{(Optional) A vector of length $n$, specifying the breaks for the x-axis}

\item{ybreaks}{(Optional) A vector of length $n$, specifying the breaks for the y-axis}

\item{xlabels}{(Optional) A vector of length $n$, specifying the labels for the x-axis}

\item{ylabels}{(Optional) A vector of length $n$, specifying the labels for the y-axis}

\item{int}{If TRUE, make the plot interactive using `ggplotly` and `layoutpub`. Default is FALSE.}

\item{tooltip}{If `int=TRUE` then this `tooltip` argument is passed to `ggplotly`. If `int=FALSE`, this is ignored. If left blank, this default is 'all', which is the default for `ggplotly`.}

\item{subtitle}{(Same as in `layoutpub`.) Is there a subtitle? Default is FALSE. This affects spacing. Ignored if `int=FALSE`.}

\item{caption}{(Same as in `layoutpub`.) Is there a caption? Default is FALSE. This affects spacing. Ignored if `int=FALSE`.}

\item{legend.rows}{(Same as in `layoutpub`.) How many rows are needed for the legend? Default is 0. This affects spacing. Ignored if `int=FALSE`.}

\item{...}{Other arguments passed to `theme_pub`}
}
\description{
This function adds theme_pub to a ggplot and changes other non-theme related options like `scale_*` and `coord_*`.
}
