% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{pubdarkred}
\alias{pubdarkred}
\title{Default dark red used by `pubtheme`}
\format{
An object of class \code{character} of length 1.
}
\usage{
pubdarkred
}
\description{
Default dark red used by `pubtheme`
}
\keyword{datasets}
