% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{pubgradgraylight}
\alias{pubgradgraylight}
\title{A color used in `pubtheme`}
\format{
An object of class \code{character} of length 1.
}
\usage{
pubgradgraylight
}
\description{
A color used in `pubtheme`
}
\keyword{datasets}
