% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{publightred}
\alias{publightred}
\title{Default light red used by `pubtheme`}
\format{
An object of class \code{character} of length 1.
}
\usage{
publightred
}
\description{
Default light red used by `pubtheme`
}
\keyword{datasets}
