% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{pubred}
\alias{pubred}
\title{Default red used by `pubtheme`}
\format{
An object of class \code{character} of length 1.
}
\usage{
pubred
}
\description{
Default red used by `pubtheme`
}
\keyword{datasets}
