% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.ggplot.defaults.r
\name{restore.ggplot.defaults}
\alias{restore.ggplot.defaults}
\title{Restore ggplot defaults}
\usage{
restore.ggplot.defaults()
}
\description{
A function for restoring ggplot defaults that were changed by the pubtheme package.
}
\examples{

## Load required packages
library(tidyverse)
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + 
    geom_point() ## default discrete colors (red, green, blue)

## pubtheme discrete colors (snred, snblue, snmediumgray)
library(pubtheme)
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + 
    geom_point() + 
    theme_pub()

## Still pubtheme discrete colors (snred, snblue, snmediumgray), even though theme_pub isn't used
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + geom_point()

## Restore defaults. Now the original colors are used again.
restore.ggplot.defaults()
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + geom_point()
ggplot(mtcars, aes(x=wt, y=mpg, color=          cyl )) + geom_point()
}
