% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse2_trans.r
\name{reverse2_trans}
\alias{reverse2_trans}
\title{Function for reverse date axes}
\usage{
reverse2_trans()
}
\description{
Copied from https://github.com/tidyverse/ggplot2/issues/4014
Use this with scale_y_continuous below (and use _continuous instead of _date)
}
