% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.R
\name{theme_pub}
\alias{theme_pub}
\title{A ggplot theme}
\usage{
theme_pub(
  type = "line",
  base_size = 36/3,
  base_family = "sans",
  base_line_size = base_size * 0.35/36 * 3,
  base_rect_size = base_size * 0.35/36,
  facet = F,
  colors = "default"
)
}
\arguments{
\item{type}{Text indicating the type of plot the theme is being used for.  Supported types are currently 'line', 'bar', 'hist', 'grid', 'scatter' which can be used for plots with `geom_line`, `geom_bar`, `geo_histogram`, `geom_tile`, and `geom_point`.  Other plot types will be added later.}

\item{base_size}{base font size, given in pts. For saving images, 36 is required. For viewing in RStudio, something much smaller like 36/3 = 12 (the default) is recommended.}

\item{base_family}{base font family.  The default is `'sans'`, which is available cross-platform by default.}

\item{base_line_size}{base size for line elements. Default is `base_size*.35/36`, and should typically not be changed.}

\item{base_rect_size}{base size for rect elements. Default is `base_size*.35/36`, and should typically not be changed.}

\item{facet}{Indicates whether or not `facet_wrap` or `facet_grid` are being used for this plot.  The default is `facet=FALSE`.}

\item{colors}{Choose the color palette. The default, colors='default", is reds, blues and grays commonly used in journalism.  'cb14' is a colorblind friendly palette with 14 colors. Choosing 'yourorgname' will use specific reds, blues, and grays from the organizations color palette, if they have been implement.}
}
\value{
When used in conjunction with ggplot, it returns a plot formatted using the sn theme.
}
\description{
A ggplot theme for making publication quality visualizations with options for organization-specific colors and logos.
}
\examples{
#See `https://github.com/bmacGTPM/pubtheme` for examples.
}
