% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.text.r
\name{tooltip.text}
\alias{tooltip.text}
\title{Function for creating HTML code for a formatted tooltip}
\usage{
tooltip.text(...)
}
\arguments{
\item{...}{Unquoted names of columns to include in the tooltip.}
}
\value{
HTML code for a tooltip where the first line is bold.
}
\description{
A function for generating HTML code to be used to create a formatted tooltip. The tooltip will have one line per column supplied to the function. The first column given will be listed in the first line and will be bolded.
}
\examples{
#See `https://github.com/bmacGTPM/pubtheme` for examples.
}
